+++ modules/timeline-layers/src/index.ts
@@ -6,3 +6,8 @@ export type {HorizonGraphLayerProps} from './layers/horizon-graph-layer/horizon-
 export {HorizonGraphLayer} from './layers/horizon-graph-layer/horizon-graph-layer';
 export type {MultiHorizonGraphLayerProps} from './layers/horizon-graph-layer/multi-horizon-graph-layer';
 export {MultiHorizonGraphLayer} from './layers/horizon-graph-layer/multi-horizon-graph-layer';
+export type {TimeAxisLayerProps} from './layers/time-axis-layer';
+export {TimeAxisLayer} from './layers/time-axis-layer';
+export type {VerticalGridLayerProps} from './layers/vertical-grid-layer';
+export {VerticalGridLayer} from './layers/vertical-grid-layer';
+export {formatTimeMs, formatTimeRangeMs} from './utils/format-utils';